Improve layering
